+++ lib/main.dart
@@ -24,17 +24,20 @@ part 'widget/player_info.dart';
 part 'widget/game_history.dart';
 part 'widget/bidding.dart';
 part 'widget/gifting.dart';
+part 'widget/round_end.dart';
 part 'widget/common.dart';
 part 'widget/selection_board.dart';
 part 'widget/team_picker.dart';
 part 'widget/decision.dart';
+part 'widget/heist_end.dart';
 part 'middleware/middleware.dart';
 part 'middleware/room_middleware.dart';
 part 'middleware/game_middleware.dart';
 part 'middleware/bidding_middleware.dart';
 part 'middleware/gifting_middleware.dart';
+part 'middleware/round_end_middleware.dart';
 part 'middleware/team_picker_middleware.dart';
-part 'middleware/decision_middleware.dart';
+part 'middleware/heist_middleware.dart';
 part 'reducers/heist_reducers.dart';
 part 'reducers/player_reducers.dart';
 part 'reducers/reducers.dart';
@@ -49,7 +52,7 @@ part 'selectors/setup_selectors.dart';
 part 'selectors/team_picker_selectors.dart';
 part 'selectors/bidding_selectors.dart';
 part 'selectors/gifting_selectors.dart';
-part 'selectors/decision_selectors.dart';
+part 'selectors/heist_selectors.dart';
 part 'state.dart';
 part 'role.dart';
 part 'heist_definitions.dart';
